Updates from master

+++ klone/klog.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2005, 2006 by KoanLogic s.r.l. <http://www.koanlogic.com>
+ * Copyright (c) 2005-2012 by KoanLogic s.r.l. <http://www.koanlogic.com>
  * All rights reserved.
  *
  * This file is part of KLone, and as such it is subject to the license stated
@@ -38,10 +38,10 @@ struct klog_args_s
     int type;           /* one of KLOG_TYPEs */
     char *ident;        /* string prepended to each log msg */
     int threshold;      /* filter log msgs lower than this level */
-    size_t mlimit;      /* max number of log messages (memory) */
+    int mlimit;         /* max number of log messages (memory) */
     char *fbasename;    /* basename of log files (postfix varies) */
-    size_t fsplits;     /* number of split files (file) */
-    size_t flimit;      /* number of log msgs per file (file) */
+    int fsplits;        /* number of split files (file) */
+    int flimit;         /* number of log msgs per file (file) */
     int soptions;       /* log options (syslog) */
     int sfacility;      /* default facility (syslog's LOG_LOCAL[0-7]) */
 #define KLOG_FACILITY_UNKNOWN   -1
